Move open_state to state submodule and resolve circular import (#1663)
