+++ asreview/__init__.py
@@ -18,7 +18,7 @@ from asreview.data.loader import load_dataset
 from asreview.data.utils import list_readers
 from asreview.data.utils import list_writers
 from asreview.project import Project
-from asreview.project import open_state
+from asreview.state.contextmanager import open_state
 from asreview.utils import asreview_path
 from asreview.utils import get_data_home
