+++ asreview/simulation/cli.py
@@ -34,9 +34,9 @@ from asreview.models.feature_extraction import get_feature_model
 from asreview.models.query import get_query_model
 from asreview.project import Project
 from asreview.project import ProjectExistsError
-from asreview.project import open_state
 from asreview.settings import ASReviewSettings
 from asreview.simulation import Simulate
+from asreview.state.contextmanager import open_state
 from asreview.types import type_n_queries
 from asreview.utils import format_to_str
 from asreview.utils import get_random_state
