+++ asreview/state/cli.py
@@ -18,7 +18,7 @@ from pathlib import Path
 import pandas as pd
 
 from asreview.project import get_project_path
-from asreview.project import open_state
+from asreview.state.contextmanager import open_state
 
 
 def _parse_state_inspect_args():
